Remove id from construct in creature model.
I do this because in other models when you have an
id in the construct it means that you are asking for
something with this id. When the id was in the construct
the id was the id of the new creature. So I did this
because I wanted the models to be consistent.